backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
anthropic==0.18.1
supabase==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
apscheduler==3.10.4
python-multipart==0.0.9
slowapi==0.1.9
